feat: Add UserAnswer option relationship
The code changes in the `schema.dbml` file add a new relationship between the `UserAnswer` and `Option` models. This relationship is defined by adding a foreign key `option_id` to the `UserAnswer` table, referencing the `id` column of the `Option` table. This change allows for tracking the selected option for each user answer.

Recent user commits:
- feat: Modify response format for login route
- feat: Update login response format
- feat: Update login response format
- feat: Update login response format

Recent repository commits:
- feat: Modify response format for login route
- feat: Update login response format
- feat: Update login response format
- feat: Update login response format
- chore: Update UserModel to match schema changes
- chore: Add prisma-dbml-generator as dev dependency and update schema.prisma
- update user model