feat: Add category field to quizDetail query in UserController
This commit adds the `category` field to the `quizDetail` query in the `UserController.js` file. The `category` field is set to `true` in the Prisma select statement, allowing the associated category information to be included in the response. This change enhances the functionality of the application by providing the category information along with the quiz details.

Code changes:
- Add `category` field to the `quizDetail` query in `UserController.js`