feat: Add API documentation link to README.md
This commit adds a link to the API documentation in the README.md file. The link directs users to the documentation page hosted at [https://docs.quizlogy.xyz](https://docs.quizlogy.xyz). This addition improves the accessibility and usability of the application by providing easy access to the API documentation.

Code changes:
- Update the README.md file to include the API documentation link